services/indexer/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
qdrant-client>=1.14.0
python-dotenv>=1.0.0
pypdf>=5.9.0
python-multipart>=0.0.20
unstructured>=0.15.0
unstructured[md]>=0.15.0
unstructured[html]>=0.15.0
ruff>=0.9.0
black>=24.12.0
isort>=5.13.0
